Add spinner while login is in progress.
