+++ client/src/actions/types.js
@@ -41,6 +41,7 @@ export const VISITS_LOADING = 'VISITS_LOADING';
 export const USER_LOADING = "USER_LOADING";
 export const USER_LOADED = "USER_LOADED";
 export const AUTH_ERROR = "AUTH_ERROR";
+export const LOGIN = "LOGIN"
 export const LOGIN_SUCCESS = "LOGIN_SUCCESS";
 export const LOGIN_FAIL = "LOGIN_FAIL";
 export const LOGIN_CANCEL = "LOGIN_CANCEL"
