+++ client/src/locales/en.js
@@ -153,6 +153,7 @@ const locale = {
         "admin": "Administration",
         "brand": "Visitor Use Management Tool",
         "cancel": "Cancel",
+        "checkingCredentials": "Checking credentials",
         "continue": "Continue",
         "home": "Home",
         "language": "Language",
